+++ pages/components/SelectedCard.module.css
