+++ pages/containers/PokemonsContainer.module.css
@@ -12,4 +12,11 @@
 .cardcontainer {
   float: left;
   margin: 0 auto;
+  display: grid;
+  grid-template-columns: 1fr auto;
+}
+
+.carditem {
+  float: left;
+  margin: 0 auto;
 }
